refactor: split task lifecycle helpers
